addapted and corrected older function and added some

+++ utilities/header/dijkstra.h
@@ -3,6 +3,6 @@
 
 #include "../../structures/header/graph_board.h"
 
-int *dijkstra(board_t *board, board_t *my_board, int node_a, int node_b, float proportion, int (*priority_calculation)(board_t *board, int node_a, int node_b, float p));
+int *dijkstra(board_t *board, board_t *my_board, int node_a, int node_b, float proportion, float (*priority_calculation)(board_t *board, int node_a, int node_b, float p));
 
 #endif
